install: Ensure all Azure CLI output is in JSON
[ cherry-picked from cilium/cilium-cli repository ]

This is useful for users who have configured their Azure CLI ("az") to
output under a different format such as "table", instead of "json",
which this tool makes the assumption of.

Example failure:

```
$ cilium install --version v1.9.1 --azure-resource-group chris
🔮 Auto-detected Kubernetes kind: AKS
🔮 Auto-detected cluster name: chris
🔮 Auto-detected IPAM mode: azure
🔮 Auto-detected datapath mode: azure
🚀 Creating service principal for Cilium operator...
Error: unable to unmarshal az output: invalid character 'A' looking for beginning of value
```

This commit fixes this by forcing that all "az" invocations run with
"--output json".

Signed-off-by: Chris Tarazi <chris@isovalent.com>