hubble/clustermesh: Add --create-ca flag to create CA if needeed
[ cherry-picked from cilium/cilium-cli repository ]

The CA is typically created at install time. Allow creating it on the
fly if needed. We don't do this automatically as the common case will be
that the CA already exists as sharing between clusters is desirable. The
default behavior should thus be to fail if it does not exist.

Signed-off-by: Thomas Graf <thomas@cilium.io>