version: Detect special running Cilium versions
[ cherry-picked from cilium/cilium-cli repository ]

When cilium is installed with 'cilium install --version="-ci:1234"'
this version string is added to the standard image names. For the
Cilium agent this becomes "quay.io/cilium/cilium-ci:1234". Detect this
when decoding version from a running Cilium pod by finding the part
after the dash in the image reference, so that the detected version
becomes "-ci:1234". This allows hubble to be enabled without passing
--hubble-version parameter explicitly in this case.

Signed-off-by: Jarno Rajahalme <jarno@isovalent.com>