connectivity: Remove nodeport checks from policy test
[ cherry-picked from cilium/cilium-cli repository ]

These tests will always fail due to [1]. The nodeport checks are kept
for the no-policies case.

Only if the following are true:
  * Direct routing
  * kube-proxy-replacement (KPR) is enabled
  * DSR enabled
then it's possible to work around [1] above.

For now, remove these tests as I think the trade-off comes down to, do
we believe we get better signal for issues when:
  a) we disallow `world`, but only test nodeport without policy
  b) keep `world` and keep testing nodeport with policy

Given the previous commit, we are moving forward with (a) because
NodePort is intended for N-S traffic hence the trade-off to skip these
checks gives us more fine-grained datapath testing.

[1]: https://github.com/cilium/cilium/issues/17144

Signed-off-by: Chris Tarazi <chris@isovalent.com>