workflows: fix AKS workflow not using the right CLI pod
[ cherry-picked from cilium/cilium-cli repository ]

The AKS workflow, like the external workload workflow, loads a shell
script to execute via a Helm chart twice: first for installing Cilium,
then for the connectivity test itself.

The Helm chart takes in additional parameters to specify which version
of the CLI should be loaded, and these were missing on the test part of
the AKS workflow, resuling in the CLI being loaded off `latest` and not
from the PR itself.

Fixes: f6cc7c6c00b7dbd25119f094c2db62f2b1b68b5f

Signed-off-by: Nicolas Busseneau <nicolas@isovalent.com>