add user-defined options into Cilium ConfigMap
[ cherry-picked from cilium/cilium-cli repository ]

In order to keep the cilium-cli options set by the user we should store
it as part of a special key in Cilium's ConfigMap. This will allow
cilium-cli to know the intent from the user for the current cluster
installation without requiring to set helm flags every time the user
interacts with the cluster via cilium-cli.

Signed-off-by: André Martins <andre@cilium.io>