.github: explicitly set Cilium version in workflows
[ cherry-picked from cilium/cilium-cli repository ]

Currently, the workflows will always install the Cilium version
specified in defaults.Version which is currently set to 1.11.4.

To ensure compatibility of changes in cilium-cli we might want to run CI
with a different Cilium version, e.g. an rc release. This change allows
specifying cilium_version in the workflows to test against a different
version.

Signed-off-by: Tobias Klauser <tobias@cilium.io>