connectivity: Add --assume-cilium-version flag
[ cherry-picked from cilium/cilium-cli repository ]

The connectivity tests run depend on the Cilium version detected. In
some cases (e.g. when run against Isovalent Cilium Enterprise) the
Cilium version is not detected correctly and the connectivity tests fall
back to assuming the latest Cilium version. These tests may fail if the
actual Cilium version is too old.

This commit adds an --assume-cilium-version flag to cilium connectivity
test to allow the version to be set explicitly, rather than falling back
to the latest by default. This is useful for CI when running the
connectivity tests against older versions of Cilium.

Signed-off-by: Tom Payne <tom@isovalent.com>