connectivity: Add --datapath for DP e2e conformance suite
[ cherry-picked from cilium/cilium-cli repository ]

It adds a single test case to test the N/S LB functionality by sending a
request from a node w/o Cilium to a NodePort service.

For now, the test case is protected by the flag "--datapath". In the
future we might drop the flag. I.e., the test cases will be part of the
regular connectivity suite.

To run the suite:

    cd $CILIUM_PATH && ./contrib/scripts/kind.sh "" 3 "" "" none
    IP=$(docker inspect kind-control-plane -f \
        '{{ .NetworkSettings.Networks.kind.IPAddress }}')
    cilium install \
        --nodes-without-cilium=kind-worker3 \
        --kube-proxy-replacement=strict \
        --helm-set-string="k8sServiceHost=${IP},k8sServicePort=6443"
    ./cilium connectivity test --datapath

Signed-off-by: Martynas Pumputis <m@lambda.lt>