connectivity: Add FlowFilters() method for TestPeer
[ cherry-picked from cilium/cilium-cli repository ]

The method is needed for flow validations when one of a pod does not
belong to a Cilium cluster, and thus that pod name cannot be used in
validating flows.

For now we extended only Pod to distinguish between the flow validations
cases we already had, and the aforementioned one.

This will be used when validating N/S LB flows.

Signed-off-by: Martynas Pumputis <m@lambda.lt>