external-workloads: explicitly configure ClusterID/Name
[ cherry-picked from cilium/cilium-cli repository ]

Currently, when the Cilium agent running on a VM joins an existing
cluster, it automatically infers the corresponding ClusterID and
ClusterName. Yet, this introduces a dependency order, because all
consumers of that field need to be initialized after the connection
to the remote cluster, which is especially problematic as more
components get converted to cells. Additionally, this may also
lead to possible user confusion, as the corresponding parameter
is reported as not initialized in the logs.

Instead, let's explicitly configure these fields in the
installation script, so that the Cilium agent can interpret
them normally.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>