kind: Set conn-disrupt-test-setup
[ cherry-picked from cilium/cilium-cli repository ]

Michi reported that reverting [1] fix the Kind CI job flake, and that
they were due to too many flow events generated by the test-conn-disrupt
pods.

As we don't care about the sensitivity of those tests in
cilium/cilium-cli, set the interval to 100ms to reduce the events count
(and thus, the test runtime).

[1]: https://github.com/cilium/cilium-cli/pull/2156

Reported-by: Michi Mutsuzaki <michi@isovalent.com>
Signed-off-by: Martynas Pumputis <m@lambda.lt>