bgp: routes CLI - update min. cilium version, use `cilium` binary
[ cherry-picked from cilium/cilium-cli repository ]

As the necessary cilium-dbg functionality for retrieving BGP routes
was backported to cilium v1.14.6 (https://github.com/cilium/cilium/pull/30205),
we have to use `cilium` binary instead of `cilium-dbg` to fetch routes
from cilium pods, as the rename from `cilium` to `cilium-dbg` happened in 1.15.

This also updates the minimal cilium version for this command.

Signed-off-by: Rastislav Szabo <rastislav.szabo@isovalent.com>