features: improve tunnel feature determination
[ cherry-picked from cilium/cilium-cli repository ]

Let's improve the consistency in the determination of the tunnel feature,
and ensure that also in case of native routing mode we retrieve the correct
tunnel protocol. Indeed, certain features (e.g., egress gateway) still rely
on the configured tunnel protocol even if the main routing mode is native.
While being there, let's also increase the test coverage of this logic.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>