action: introduce the NewGenericAction method
[ cherry-picked from cilium/cilium-cli repository ]

NewGenericAction mimics the behavior of the already existing NewAction
method, but is intended to be used for assertions that are not strictly
connectivity checks (e.g., asserting the absence of log errors over
multiple pods), hence not associated with any execution pod nor network
target). Yet, it still provides improved progress and error reporting
compared to just a bare scenario.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>