Add source IP visibility info to Ingress and Gateway API docs
[ upstream commit 5e981cce0209f8838de91f0c55f5d981488adbe8 ]
[ backporter's note: conflicts under Documentation/network/servicemesh/:
  1. host-network-mode.rst: 1.15 doesn't have this file.
  2. troubleshooting.rst: 1.15 uses `-` rather than `=`.
  3. ingress.rst: 1.15 doesn't have `Ingress Path Types and
     Precedence` section. ]

This commit adds information about how best to get access to
the source IP address for Ingress and Gateway API configs.

As part of this, break out the Reference section recently
added to the Ingress config, and bring it into the Gateway API
docs as well (for all the areas where the same things apply).

Signed-off-by: Nick Young <ynick@cisco.com>
Signed-off-by: gray <gray.liang@isovalent.com>