ctmap: remove unused valid gc filter field.
While performing gc, this field would remove entries that had
a src/dst IP that wasn't part of a defined set of 'valid' IPs.

This filter was skipped if the ValidIPs field was left nil. There
appears to be no usage of this field.

Therefore, to simplify the GC logic this commit removes that
functionality.

Signed-off-by: Tom Hadlaw <tom.hadlaw@isovalent.com>